Move entity drawing to AreaDsl for scissoring
Entity drawing is now scissored by default (for performance?), which means that we need to supply two more coordinates for the box ending extents. Instead, it's just easier to move entity drawing into the AreaDsl, which has it's own defined extents. Also, we can set a better default size for entity drawing, being the min value of the size of the area. (size 15 fit almost perfectly into a 16x16 pixel grid before, which is why we are saying 1 size unit is ~1 pixel right now).